third_party/WebKit/LayoutTests/http/tests/security/xss-DENIED-getSVGDocument-object-expected.txt

CONSOLE MESSAGE: line 14: PASS: SVGDocument access threw: 'SecurityError: Blocked a frame with origin "http://127.0.0.1:8000" from accessing a cross-origin frame.'.
